fix: do not delete primary schedule of script/flow on redeploy even if schedule wasn't loaded

+++ frontend/src/lib/components/FlowBuilder.svelte
@@ -330,7 +330,7 @@
 								requestBody: { enabled }
 							})
 						}
-					} else if (scheduleExists) {
+					} else if (scheduleExists && !$triggersCount?.primary_schedule) {
 						await ScheduleService.deleteSchedule({
 							workspace: $workspaceStore ?? '',
 							path: $pathStore
